[IMP] make screenshake depend on ball velocity, lower intensity of shake
Previously, the direction of the screenshake depended of the position of
the collistion relative to the center of the screen, and the intensity
of the shake was hardcodede.

This commit makes the direction of the screenshake depend on the
direction of the collision, and its intensity depend on the velocity of
that collision. There is now also a threshold velocity for collisions
under which they will not produce any screenshake at all.